refactor : 폰트 추가 및 필요없는 코드 제거
