+++ pages/_app.tsx
@@ -1,13 +1,12 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import Header from "../components/layout/header/Header";
-import ColorModeContextProvider from "../components/context/ColorModeContextProvider";
+import ColorModeContextProvider from "../context/ColorModeContextProvider";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <ColorModeContextProvider>
       <Header />
-
       <Component {...pageProps} />
     </ColorModeContextProvider>
   );
